Optimization.
Moved all value members of JSONValue to an unnamed union so that extra
memory isn't wasted. Changing this made it so the JSONObject, JSONArray
and JSONString have to be pointers and freed as such. Converting them
into the heap causes no impact on performance because std::vector,
std::map and std::wstring already make allocations in the heap.